Apply the single responsibility principle
Refactor code to apply the Single Responsibility Principle.

* **Add `format_checker.py`**: Add `check_formatting` function to handle formatting checks.
* **Modify `gherkin_parser.py`**: Remove `check_formatting` function.
* **Modify `main.py`**: Import `check_formatting` from `format_checker.py` instead of `gherkin_parser.py`.
* **Update `README.md`**: Include `format_checker.py` in the overview and usage instructions.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/Zeegar/FeatureFile2Fibery?shareId=XXXX-XXXX-XXXX-XXXX).
