+++ gherkin_parser.py
@@ -28,22 +28,3 @@ def parse_feature_file(file_path):
             print(f"Invalid Gherkin syntax at line {line_number}: {line}")
 
     return features
-
-
-def check_formatting(file_path):
-    with open(file_path, 'r') as file:
-        lines = file.readlines()
-
-    errors = []
-    for i, line in enumerate(lines):
-        line = line.strip()
-        if not any(line.startswith(keyword) for keyword in
-                   ['Feature:', ' ', '', '@', 'Scenario:',
-                    'Scenario Outline:', 'Developer Task:', 'Given', 'And',
-                    'When', 'Then']):
-            errors.append(f"Formatting error on line {i + 1}: {line}")
-
-    if errors:
-        return "\n".join(errors)
-    else:
-        return "Formatting Ok"
